fixed buzz button not showing
