+++ src/components/pages/Feeds/index.js
@@ -1,5 +1,5 @@
-import React, { useEffect, useCallback, useState } from 'react'
-import { CreateBuzzForm, InfiniteList, HelmetGenerator, BuzzFormModal } from 'components'
+import React, { useEffect, useCallback } from 'react'
+import { CreateBuzzForm, InfiniteList, HelmetGenerator } from 'components'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { pending } from 'redux-saga-thunk'
@@ -27,12 +27,9 @@ import {
   clearAccountReplies,
 } from 'store/profile/actions'
 import { clearScrollIndex, clearRefreshRouteStatus} from 'store/interface/actions'
-import { anchorTop, useWindowDimensions } from 'services/helper'
+import { anchorTop } from 'services/helper'
 import { isMobile } from 'react-device-detect'
-import { Link, useHistory, useLocation } from 'react-router-dom'
-import { setBuzzModalStatus } from 'store/interface/actions'
-import { Fab } from '@material-ui/core'
-import BuzzIcon from 'components/elements/Icons/BuzzIcon'
+import { Link } from 'react-router-dom'
 
 
 const useStyles = createUseStyles(theme => ({
@@ -44,16 +41,6 @@ const useStyles = createUseStyles(theme => ({
     },
   },
 }))
-
-const floatStyle = {
-  margin: 0,
-  top: 'auto',
-  bottom: 20,
-  left: 'auto',
-  position: 'fixed',
-  zIndex: 500,
-  backgroundColor: '#e61c34',
-}
 
 const Feeds = React.memo((props) => {
   const {
@@ -85,13 +72,6 @@ const Feeds = React.memo((props) => {
     clearRefreshRouteStatus,
   } = props
   const classes = useStyles()
-  const location = useLocation()
-  const history = useHistory()
-  const { pathname } = location
-  const [open, setOpen] = useState(false)
-  const [minify, setMinify] = useState(false)
-  const { width } = useWindowDimensions()
-  const isBuzzIntent = pathname.match(/^\/intent\/buzz/)
 
   useEffect(() => {
     setPageFrom('home')
@@ -129,39 +109,11 @@ const Feeds = React.memo((props) => {
     // eslint-disable-next-line
   }, [refreshRouteStatus])
 
-  useEffect(() => {
-    if(width >= 1366) {
-      setMinify(false)
-    } else if(width >= 1026 && width < 1366) {
-      setMinify(true)
-    } else if(width >=706 && width < 1026) {
-      setMinify(true)
-    } else {
-      setMinify(true)
-    }
-  }, [width])
-
   const loadMorePosts = useCallback(() => {
     const { permlink, author } = last
     getHomePostsRequest(permlink, author)
     // eslint-disable-next-line
   }, [last])
-
-  const handleOpenBuzzModal = () => {
-    setOpen(true)
-  }
-
-  const onHide = () => {
-    setBuzzModalStatus(false)
-    setOpen(false)
-    if (isBuzzIntent) {
-      history.push('/')
-    }
-  }
-
-  const floatingButtonStyle = {
-    marginLeft: width > 1026 && 530, right: width < 1026 && 30,
-  }
   
   return (
     <React.Fragment>
@@ -180,25 +132,16 @@ const Feeds = React.memo((props) => {
         </React.Fragment>
       )}
       <InfiniteList loading={loading} items={items} onScroll={loadMorePosts} />
-      {minify && (
-        <Fab onClick={handleOpenBuzzModal} size="medium" color="secondary" aria-label="add" style={{...floatStyle, ...floatingButtonStyle}}>
-          <BuzzIcon />
-        </Fab>
-      )}
-      <BuzzFormModal show={open} onHide={onHide} />
     </React.Fragment>
   )
 })
 
 const mapStateToProps = (state) => ({
-  buzzModalStatus: state.interfaces.get('buzzModalStatus'),
   loading: pending(state, 'GET_HOME_POSTS_REQUEST'),
   isHomeVisited: state.posts.get('isHomeVisited'),
   items: state.posts.get('home'),
   last: state.posts.get('lastHome'),
   refreshRouteStatus: state.interfaces.get('refreshRouteStatus'),
-  intentBuzz: state.auth.get("intentBuzz"),
-  fromIntentBuzz: state.auth.get('fromIntentBuzz'),
 })
 
 const mapDispatchToProps = (dispatch) => ({
@@ -223,7 +166,6 @@ const mapDispatchToProps = (dispatch) => ({
     clearScrollIndex,
     clearHomePosts,
     clearRefreshRouteStatus,
-    setBuzzModalStatus,
   }, dispatch),
 })
